Introduce YAML zone and restructure YAML tests
Marks production code with ILanguageYamlZone and sets up test zones, activators and components in separate namespaces.
