+++ resharper/resharper-yaml/src/ZoneMarker.cs
@@ -1,0 +1,9 @@
+using JetBrains.Application.BuildScript.Application.Zones;
+
+namespace JetBrains.ReSharper.Plugins.Yaml
+{
+  [ZoneMarker]
+  public class ZoneMarker : IRequire<ILanguageYamlZone>
+  {
+  }
+}
