+++ resharper/resharper-yaml/test/src/Yaml/ProjectModel/YamlProjectFileTypeTests.cs
