+++ resharper/resharper-yaml/test/src/Yaml/Psi/Parsing/IncrementalReparseTestBase.cs
