+++ resharper/resharper-yaml/test/src/Yaml/ZoneMarker.cs
@@ -1,0 +1,9 @@
+using JetBrains.Application.BuildScript.Application.Zones;
+
+namespace JetBrains.ReSharper.Plugins.Tests.Yaml
+{
+  [ZoneMarker]
+  public class ZoneMarker : IRequire<IYamlTestsZone>
+  {
+  }
+}
