+++ resharper/resharper-yaml/test/src/YamlTestComponents/ZoneMarker.cs
@@ -1,0 +1,12 @@
+using JetBrains.Application.BuildScript.Application.Zones;
+using JetBrains.ProjectModel;
+
+namespace JetBrains.ReSharper.Plugins.Tests.YamlTestComponents
+{
+  // Zone requirements for non-environment test components. Separate namespace to environment components to avoid
+  // adding inactive zones as requirements to environment components
+  [ZoneMarker]
+  public class ZoneMarker : IRequire<IProjectModelZone>
+  {
+  }
+}
