+++ resharper/resharper-yaml/test/src/tests.resharper-yaml.csproj
@@ -2,7 +2,7 @@
   <PropertyGroup>
     <TargetFramework>net472</TargetFramework>
     <AssemblyName>JetBrains.ReSharper.Plugins.Yaml.Tests</AssemblyName>
-    <RootNamespace>JetBrains.ReSharper.Plugins.Tests.Yaml</RootNamespace>
+    <RootNamespace>JetBrains.ReSharper.Plugins.Tests</RootNamespace>
     <LangVersion>9</LangVersion>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <JetTestProject>True</JetTestProject>
